requirements.txt
Django==4.2.9
python-dotenv==1.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
pytest==8.0.0
pytest-django==4.8.0
model-bakery==1.17.0